declaring global interfaces; script tags to top of sfcs
