+++ packages/renderer/src/components/ButtonRouterLink.vue
@@ -1,22 +1,3 @@
-<template>
-  <router-link
-    class="router-link"
-    :class="{ 'router-link-disabled': isDisabled }"
-    :to="route"
-    draggable="false"
-  >
-    <base-button-toggle
-      :is-active="isActiveRoute"
-      :is-disabled="isDisabled"
-    >
-      <template #btn-icon>
-        <slot name="icon" />
-      </template>
-      <slot />
-    </base-button-toggle>
-  </router-link>
-</template>
-
 <script setup lang="ts">
 import { computed, inject } from "vue";
 import BaseButtonToggle from "./BaseButtonToggle.vue";
@@ -40,6 +21,25 @@ const isActiveRoute = computed(() =>
   store.application.state.activeView === props.route ? true : false,
 );
 </script>
+
+<template>
+  <router-link
+    class="router-link"
+    :class="{ 'router-link-disabled': isDisabled }"
+    :to="route"
+    draggable="false"
+  >
+    <base-button-toggle
+      :is-active="isActiveRoute"
+      :is-disabled="isDisabled"
+    >
+      <template #btn-icon>
+        <slot name="icon" />
+      </template>
+      <slot />
+    </base-button-toggle>
+  </router-link>
+</template>
 
 <style scoped>
 .router-link {
