+++ packages/renderer/src/components/TheColumnProjectList.vue
@@ -1,12 +1,3 @@
-<template>
-  <base-button-click
-    :background-color="'var(--white)'"
-    @click="emitClick"
-  >
-    Create
-  </base-button-click>
-</template>
-
 <script setup lang="ts">
 import BaseButtonClick from "./BaseButtonClick.vue";
 
@@ -17,3 +8,12 @@ function emitClick(): void {
 	emit("createClick");
 }
 </script>
+
+<template>
+  <base-button-click
+    :background-color="'var(--white)'"
+    @click="emitClick"
+  >
+    Create
+  </base-button-click>
+</template>
