+++ packages/renderer/src/views/Welcome.vue
@@ -1,3 +1,18 @@
+<script setup lang="ts">
+import { ref } from "vue";
+import BaseCard from "../components/BaseCard.vue";
+import BaseCardContent from "../components/BaseCardContent.vue";
+import BaseButtonClick from "../components/BaseButtonClick.vue";
+import FormProjectCreateModal from "../components/FormProjectCreateModal.vue";
+import NotificationDot from "../components/NotificationDot.vue";
+
+const isModalActive = ref<boolean>(false);
+
+function openWindow(): void {
+	console.log("Open window in main process");
+}
+</script>
+
 <template>
   <base-card>
     <template #header>
@@ -70,18 +85,3 @@
     @close-modal="isModalActive = false"
   />
 </template>
-
-<script setup lang="ts">
-import { ref } from "vue";
-import BaseCard from "../components/BaseCard.vue";
-import BaseCardContent from "../components/BaseCardContent.vue";
-import BaseButtonClick from "../components/BaseButtonClick.vue";
-import FormProjectCreateModal from "../components/FormProjectCreateModal.vue";
-import NotificationDot from "../components/NotificationDot.vue";
-
-const isModalActive = ref<boolean>(false);
-
-function openWindow(): void {
-	console.log("Open window in main process");
-}
-</script>
