+++ tsconfig.json
@@ -9,7 +9,7 @@
     "isolatedModules": true,
     "esModuleInterop": true,
     "types": ["node"],
-    "typeRoots": ["node_modules/@types"],
+    "typeRoots": ["node_modules/@types", "./@types"],
     "lib": ["ESNext"]
   }
 }
